Refactor: Update paths

+++ Microshoppy.Sales/Microshoppy.Sales/src/SalesProduct.cs
@@ -10,6 +10,6 @@ namespace Microshoppy.Sales
 
 		public decimal NetPrice { get; set; }
 
-		public int TaxPercentage { get; set; }
+		public decimal TaxPercentage { get; set; }
 	}
 }
